requirements (1).txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.10.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.60.0
timm>=0.9.0
pytesseract>=0.3.0
opencv-python>=4.5.0

# Optional for better performance
accelerate>=0.20.0
tokenizers>=0.13.0

# For development and debugging
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0

# For Gradio interface
gradio>=4.0.0
plotly>=5.0.0
pandas>=1.3.0
kaleido>=0.2.1 